feat(core): extract db/ into core/ kernel (Phase C)
Move postgres.ts, postgres-adapter.ts, app-db.ts, app-db-router.ts,
timescale.ts, test-db-helper.ts, postgres.test.ts, and both migration
directories (postgres-migrations/, timescale-migrations/) to
backend/src/core/db/. Migration __dirname resolution preserved via
import.meta.url (relative path unchanged). Create re-export shims.

Part of #706
