+++ backend/src/core/db/postgres-migrations/002_settings.sql
