+++ backend/src/core/db/postgres-migrations/003_insights.sql
