+++ backend/src/core/db/postgres-migrations/007_investigations.sql
