+++ backend/src/core/db/postgres-migrations/019_image_staleness.sql
