+++ backend/src/core/db/postgres-migrations/021_kpi_snapshots.sql
