+++ backend/src/core/db/postgres-migrations/022_seed_default_prompts.sql
