+++ backend/src/core/db/postgres-migrations/023_harbor_vulnerabilities.sql
