+++ backend/src/core/db/postgres-migrations/024_prompt_versions.sql
