+++ backend/src/core/db/postgres-migrations/025_command_palette_filter_action.sql
