+++ backend/src/core/db/timescale-migrations/001_metrics_hypertable.sql
